untested CS changes
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@2041 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcaqgui2dfactory/gpselement.h
@@ -81,9 +81,9 @@ public:
 
     // inherited from IKnowsPlatformPosition2d
     // Access to coordinates in GUI coordinate system [m] [m] [rad]
-    virtual float x() const { return x_; }
-    virtual float y() const { return y_; }
-    virtual float theta() const { return theta_; }
+    virtual double x() const { return x_; }
+    virtual double y() const { return y_; }
+    virtual double theta() const { return theta_; }
     virtual int platformKnowledgeReliability() const { return 7; }
